(split from main repo) Try to find ssh.exe in git installation directory
GIT_SSH environment variable is not always set. As a result, ssh is not found,
resulting in various failures. On of them is that the gerrit plugin fails to
install the commit-msg hook, with a very unuseful message.
Of course, there is still a possibility to set this environemnt variable, but
this is an extra step.

The change adds a try to retrieve ssh.exe from git installation directory,
when no environment variable is found.