(split from main repo) Substring review and extension methods
We do a lot of string manipulation. Finding indexes and obtaining
substrings can lead to verbose and confusing code, and bugs.

This change introduces a few extension methods for working with
strings that leads to clearer code in several places.

- RemovePrefix
- RemoveSuffix
- SubstringUntil
- SubstringUntilLast
- SubstringAfter
- SubstringAfterLast

It replaces some older versions of these (SkipStr, SubstringUntil) with
versions that don't special case null values.