(split from main repo) Settings: Prevent the multiple creation of controls
because `GetSettings()` (where controls are created) is call 3 times
that ends with that the controls displayed are not the one
that are the ones set in `CustomControl` property.

And that can cause problems if we manipulate `CustomControl`in the setting page
to get (until the setting is saved) or set a value.

We **must** keep the strategy that `GetSettings()` return every time new controls
because the previous instances of the controls are disposed when the settings form is closed.

These changes ensure that the `GetSettings()` method is called only **1** time.

The 2 other times (use of `.Any()`) has been replaced by a `HasSettings` property.